NvFetcher

Description

The main module of nvfetcher. If you want to create CLI program with it, it's enough to import only this module.

Example:

module Main where

import NvFetcher

main :: IO ()
main = defaultMain defaultArgs packageSet

packageSet :: PackageSet ()
packageSet = do
  define $ package "feeluown-core" fromPypi "feeluown"
  define $ package "qliveplayer" fromGitHub (IsoaSFlus, QLivePlayer)

You can find more examples of packages in Main_example.hs.

Running the created program:

main -- abbreviation of main build

main build -- build nix sources expr from given packageSet

main clean -- delete .shake dir and generated nix file

main -j -- build with parallelism

All shake options are inherited.

Documentation

Re-export DSL things

module NvFetcher.PackageSet

module NvFetcher.Types

nvfetcherRules :: Rules ()

Rules of nvfetcher

generateNixSources :: FilePath -> [Package] -> Action ()

Main action, given a set of packages, generating nix sources expr in a file

data Args

Arguments for running nvfetcher

Constructors

Args

Fields

argShakeOptions :: ShakeOptions -> ShakeOptions

tweak shake options

argOutputFilePath :: FilePath

Output file path

argRules :: Rules ()

Custom rules

argActionAfterBuild :: Action ()

Action run after build rule

argActionAfterClean :: Action ()

Action run after clean rule

defaultArgs :: Args

Default arguments of defaultMain

Output file path is sources.nix.

defaultMain :: Args -> PackageSet () -> IO ()

Entry point of nvfetcher

defaultMainWith :: [OptDescr (Either String a)] -> ([a] -> IO (Args, PackageSet ())) -> IO ()

Like defaultMain but allows to define custom cli flags

data VersionChange

Version change of a package

>>> VersionChange "foo" Nothing "2.3.3"
foo: ∅ → 2.3.3

>>> VersionChange "bar" (Just "2.2.2") "2.3.3"
bar: 2.2.2 → 2.3.3

Constructors

VersionChange

Fields

vcName :: PackageName

vcOld :: Maybe Version

vcNew :: Version

getVersionChanges :: Action [VersionChange]

Get version changes since the last run, relying on shake database.

Use this function in argActionAfterBuild to produce external changelog