NvFetcher.ExtractSrc

Description

This module provides function that extracs files contents from package sources. It uses IFD under the hood, pulling textual files from source drv. Because we use nix-instantiate to build drv, so nixpkgs (NIX_PATH) is required.

Types

data ExtractSrcQ

Extract file contents from package source e.g. Cargo.lock

Constructors

ExtractSrcQ (NixFetcher Fetched) (NonEmpty FilePath)

Rules

extractSrcRule :: Rules ()

Rules of extract source

Functions

extractSrc

Arguments

:: NixFetcher Fetched
prefetched source

-> FilePath
relative file path to extract

-> Action (HashMap FilePath Text)

Run extract source

extractSrcs

Arguments

:: NixFetcher Fetched
prefetched source

-> NonEmpty FilePath
relative file paths to extract

-> Action (HashMap FilePath Text)

Run extract source with many sources