NvFetcher.GetGitCommitDate

Description

This module provides mechanisms for obtaining the git commit date. The cloned repo will not be preserved.

Types

newtype GitDateFormat

strftime format Defaults to %Y-%m-%d

Constructors

GitDateFormat (Maybe Text)

data GetGitCommitDate

Get the commit date by using shallow clone

_gformat is in. Note: Requires git >= 2.5

Constructors

GetGitCommitDate

Fields

_gurl :: Text

_grev :: Text

_gformat :: (GitDateFormat, GitTimeZone)

newtype GitTimeZone

Defaults to commit's time zone. When set to local, current local time zone is used. Only used in GetGitCommitDate.

Constructors

GitTimeZone (Maybe Text)

Rules

getGitCommitDateRule :: Rules ()

Functions

getGitCommitDate :: Text -> Text -> (GitDateFormat, GitTimeZone) -> Action Text