NvFetcher.NixExpr

Description

This module contains a type class ToNixExpr and some its instances associated with either Haskell primitive types or our NvFetcher.Types.

Documentation

type NixExpr = Text

Nix expression

class ToNixExpr a where

Types can be converted into nix expr

Methods

toNixExpr :: a -> NixExpr

fetcherToDrv :: NixFetcher Fetched -> Text -> NixExpr

Create a trivial drv that extracts the source from a fetcher TODO: Avoid using NIX_PATH