NvFetcher.NixFetcher

Description

NixFetcher is used to describe how to fetch package sources.

There are three types of fetchers overall:

FetchGit -- nix-prefetch fetchgit

FetchGitHub -- nix-prefetch fetchFromGitHub

FetchUrl -- nix-prefetch fetchurl

As you can see the type signature of prefetch: a fetcher will be filled with the fetch result (hash) after the prefetch.

Types

data RunFetch

The input of prefetch rule

Constructors

RunFetch ForceFetch (NixFetcher Fresh)

data ForceFetch

Whether to cache the fetched sha256

ForceFetch indicates alwaysRerun the fetcher rule

Constructors

ForceFetch

NoForceFetch

data NixFetcher (k :: FetchStatus)

If the package is prefetched, then we can obtain the SHA256

Constructors

FetchGit

Fields

_furl :: Text

_rev :: Version

_deepClone :: Bool

_fetchSubmodules :: Bool

_leaveDotGit :: Bool

_name :: Maybe Text

_sha256 :: FetchResult Checksum k

FetchGitHub

Fields

_fowner :: Text

_frepo :: Text

_rev :: Version

_deepClone :: Bool

_fetchSubmodules :: Bool

_leaveDotGit :: Bool

_name :: Maybe Text

_sha256 :: FetchResult Checksum k

FetchUrl

Fields

_furl :: Text

_name :: Maybe Text

_sha256 :: FetchResult Checksum k

FetchTarball

Fields

_furl :: Text

_sha256 :: FetchResult Checksum k

FetchDocker

Fields

_imageName :: Text

_imageTag :: Text

_imageDigest :: FetchResult ContainerDigest k

_sha256 :: FetchResult Checksum k

_fos :: Maybe Text

_farch :: Maybe Text

_finalImageName :: Maybe Text

_finalImageTag :: Maybe Text

_tlsVerify :: Maybe Bool

data FetchStatus

Fetch status

Constructors

Fresh

Fetched

type family FetchResult a (k :: FetchStatus) where ...

Prefetched fetchers hold hashes

Equations

FetchResult _ Fresh = ()

FetchResult a Fetched = a

Rules

prefetchRule :: Rules ()

Rules of nix fetcher

prefetch :: NixFetcher Fresh -> ForceFetch -> Action (NixFetcher Fetched)

Run nix fetcher

Functions

gitHubFetcher

Arguments

:: (Text, Text)
owner and repo

-> PackageFetcher

Create a fetcher from github repo

pypiFetcher :: Text -> PackageFetcher

Create a fetcher from pypi

gitHubReleaseFetcher

Arguments

:: (Text, Text)
owner and repo

-> Text
file name

-> PackageFetcher

Create a fetcher from github release

gitHubReleaseFetcher'

Arguments

:: (Text, Text)
owner and repo

-> (Version -> Text)
file name computed from version

-> PackageFetcher

Create a fetcher from github release

gitFetcher :: Text -> PackageFetcher

Create a fetcher from git url

urlFetcher :: Text -> NixFetcher Fresh

Create a fetcher from url

openVsxFetcher

Arguments

:: (Text, Text)
publisher and extension name

-> PackageFetcher

Create a fetcher from openvsx

vscodeMarketplaceFetcher

Arguments

:: (Text, Text)
publisher and extension name

-> PackageFetcher

Create a fetcher from vscode marketplace

tarballFetcher :: Text -> NixFetcher Fresh

Create a fetcher from url, using fetchTarball