NvFetcher.Nvchecker

Description

nvchecker is a program checking new versions of packages. We encode the checking process into shake build system, generating configuration of nvchecker and calling it externally. Now we call nvchecker for each VersionSource, which seems not to be efficient, but it's tolerable when running in parallel.

Meanwhile, we lose the capabilities of tracking version updates, i.e. normally nvchecker will help us maintain a list of old versions, so that we are able to know which package's version is updated in this run. Fortunately, we can reimplement this in shake, see nvcheckerRule for details.

Types

data VersionSortMethod

Constructors

ParseVersion

Vercmp

data ListOptions

Filter-like configuration for some version sources. See https://nvchecker.readthedocs.io/en/latest/usage.html#list-options for details.

Constructors

ListOptions

Fields

_includeRegex :: Maybe Text

_excludeRegex :: Maybe Text

_sortVersionKey :: Maybe VersionSortMethod

_ignored :: Maybe Text

data CheckVersion

The input of nvchecker

Constructors

CheckVersion VersionSource NvcheckerOptions

data NvcheckerOptions

Configuration available for evey version sourece. See https://nvchecker.readthedocs.io/en/latest/usage.html#global-options for details.

Constructors

NvcheckerOptions

Fields

_stripPrefix :: Maybe Text

_fromPattern :: Maybe Text

_toPattern :: Maybe Text

data VersionSource

Upstream version source for nvchecker to check

Constructors

GitHubRelease

Fields

_owner :: Text

_repo :: Text

GitHubTag

Fields

_owner :: Text

_repo :: Text

_listOptions :: ListOptions

Git

Fields

_vurl :: Text

_vbranch :: Branch

Pypi

Fields

_pypi :: Text

ArchLinux

Fields

_archpkg :: Text

Aur

Fields

_aur :: Text

Manual

Fields

_manual :: Text

Repology

Fields

_repology :: Text

_repo :: Text

Webpage

Fields

_vurl :: Text

_regex :: Text

_listOptions :: ListOptions

HttpHeader

Fields

_vurl :: Text

_regex :: Text

_listOptions :: ListOptions

OpenVsx

Fields

_ovPublisher :: Text

_ovExtName :: Text

VscodeMarketplace

Fields

_vsmPublisher :: Text

_vsmExtName :: Text

Cmd

Fields

_vcmd :: Text

data NvcheckerResult

The result of nvchecker rule

Constructors

NvcheckerResult

Fields

nvNow :: Version

nvOld :: Maybe Version

last result of this nvchecker rule TODO: consider removing this field

nvStale :: Bool

stale means even nvNow comes from json file (last run) and we actually didn't run nvchecker this time. nvOld will be Nothing in this case.

Rules

nvcheckerRule :: Rules ()

Rules of nvchecker

Functions

checkVersion :: VersionSource -> NvcheckerOptions -> PackageKey -> Action NvcheckerResult

Run nvchecker given PackageKey Recording version changes and using stale version are available.

checkVersion' :: VersionSource -> NvcheckerOptions -> Action NvcheckerResult

Run nvchecker without cache