NvFetcher.PackageSet

Description

This module mainly contains two things: PackageSet and PkgDSL. NvFetcher accepts the former one -- a set of packages to produce nix sources expr; the later one is used to construct a single package.

There are many combinators for defining packages. See the documentation of define for example.

Package set

data PackageSetF f

Atomic terms of package set

type PackageSet = Free PackageSetF

Package set is a monad equipped with two capabilities:

Carry defined packages

Run shake actions

Consturct it using newPackage and embedAction

newPackage :: PackageName -> VersionSource -> PackageFetcher -> PackageSet ()

Add a package to package set

embedAction :: Action a -> PackageSet a

Lift a shake Action to package set

purePackageSet :: [Package] -> PackageSet ()

Add a list of packages into package set

runPackageSet :: PackageSet () -> Action (Set Package)

Run package set into a set of packages

Throws exception as more then one packages with the same name are defined

Package DSL

Primitives

class PkgDSL f where

A tagless final style DSL for constructing packages

Methods

new :: f PackageName -> f (Prod '[PackageName])

andThen :: f (Prod r) -> f a -> f (Prod (a ': r))

end :: (Member PackageName r, Member VersionSource r, Member PackageFetcher r) => f (Prod r) -> f ()

define :: (Member PackageName r, Member VersionSource r, Member PackageFetcher r) => PackageSet (Prod r) -> PackageSet ()

PkgDSL version of newPackage

Example:

define $ package "nvfetcher-git" sourceGit "nvfetcher" fetchGitHub ("berberman", "nvfetcher")

package :: PackageName -> PackageSet (Prod '[PackageName])

Start chaining with the name of package to define

src :: PackageSet (Prod r) -> VersionSource -> PackageSet (Prod (VersionSource ': r))

Attach version sources

fetch :: PackageSet (Prod r) -> PackageFetcher -> PackageSet (Prod (PackageFetcher ': r))

Attach fetchers

Two-in-one functions

fromGitHub :: PackageSet (Prod r) -> (Text, Text) -> PackageSet (Prod (PackageFetcher ': (VersionSource ': r)))

A synonym of fetchGitHub and sourceGitHub

fromPypi :: PackageSet (Prod r) -> Text -> PackageSet (Prod (PackageFetcher ': (VersionSource ': r)))

A synonym of fetchPypi and sourcePypi

Version sources

sourceGitHub

Arguments

:: PackageSet (Prod r)

-> (Text, Text)
owner and repo

-> PackageSet (Prod (VersionSource ': r))

This package follows the latest github release

sourceGit

Arguments

:: PackageSet (Prod r)

-> Text
git url

-> PackageSet (Prod (VersionSource ': r))

This package follows the latest git commit

sourcePypi

Arguments

:: PackageSet (Prod r)

-> Text
pypi name

-> PackageSet (Prod (VersionSource ': r))

This package follows the latest pypi release

sourceAur

Arguments

:: PackageSet (Prod r)

-> Text
package name in Aur

-> PackageSet (Prod (VersionSource ': r))

This package follows the version of an Aur package

sourceArchLinux

Arguments

:: PackageSet (Prod r)

-> Text
package name in Arch Linux repo

-> PackageSet (Prod (VersionSource ': r))

This package follows the version of an Arch Linux package

sourceManual :: PackageSet (Prod r) -> Text -> PackageSet (Prod (VersionSource ': r))

This package follows a pinned version

sourceRepology

Arguments

:: PackageSet (Prod r)

-> (Text, Text)
repology project name and repo

-> PackageSet (Prod (VersionSource ': r))

This package follows the version of a repology package

Fetchers

fetchGitHub

Arguments

:: PackageSet (Prod r)

-> (Text, Text)
owner and repo

-> PackageSet (Prod (PackageFetcher ': r))

This package is fetched from a github repo

fetchGitHubRelease

Arguments

:: PackageSet (Prod r)

-> (Text, Text, Text)
owner, repo, and file name

-> PackageSet (Prod (PackageFetcher ': r))

This package is fetched from a file in github release

fetchPypi

Arguments

:: PackageSet (Prod r)

-> Text
pypi name

-> PackageSet (Prod (PackageFetcher ': r))

This package is fetched from pypi

fetchGit

Arguments

:: PackageSet (Prod r)

-> Text
git url

-> PackageSet (Prod (PackageFetcher ': r))

This package is fetched from git

fetchUrl

Arguments

:: PackageSet (Prod r)

-> (Version -> Text)
url, given a specific version

-> PackageSet (Prod (PackageFetcher ': r))

This package is fetched from url

Miscellaneous

data Prod (r :: [Type]) where

Simple HList

Constructors

Nil :: Prod '[]

Cons :: !x -> Prod xs -> Prod (x ': xs)

class Member (a :: Type) (r :: [Type]) where

Project elements from Prod

Methods

proj :: Prod r -> a

type family NotElem (x :: Type) (xs :: [Type]) :: Constraint where ...

Constraint for producing error messages

Equations

NotElem x (x ': xs) = TypeError (ShowType x :<>: 'Text " is defined more than one times")

NotElem x (_ ': xs) = NotElem x xs

NotElem x '[] = ()

coerce :: forall (k :: RuntimeRep) (a :: TYPE k) (b :: TYPE k). Coercible a b => a -> b

The function coerce allows you to safely convert between values of types that have the same representation with no run-time overhead. In the simplest case you can use it instead of a newtype constructor, to go from the newtype's concrete type to the abstract type. But it also works in more complicated settings, e.g. converting a list of newtypes to a list of concrete types.

This function is runtime-representation polymorphic, but the RuntimeRep type argument is marked as Inferred, meaning that it is not available for visible type application. This means the typechecker will accept coerce @Int @Age 42.