NvFetcher.Types

Description

Types used in this program.

Common types

newtype Version

Package version

Constructors

Version Text

newtype SHA256

SHA 256 sum

Constructors

SHA256 Text

Nvchecker types

data VersionSource

The input of nvchecker

Constructors

GitHubRelease

Fields

owner :: Text

repo :: Text

Git

Fields

vurl :: Text

Pypi

Fields

pypi :: Text

ArchLinux

Fields

archpkg :: Text

Aur

Fields

aur :: Text

Manual

Fields

manual :: Text

Repology

Fields

repology :: Text

repo :: Text

data NvcheckerResult

The result of running nvchecker

Constructors

NvcheckerResult

Fields

nvNow :: Version

nvOld :: Maybe Version

nvchecker doesn't give this value, but shake restores it from last run

Nix fetcher types

data NixFetcher (k :: Prefetch)

If the package is prefetched, then we can obtain the SHA256

Constructors

FetchGit

Fields

furl :: Text

rev :: Version

branch :: Maybe Text

deepClone :: Bool

fetchSubmodules :: Bool

leaveDotGit :: Bool

sha256 :: PrefetchResult k

FetchUrl

Fields

furl :: Text

sha256 :: PrefetchResult k

data Prefetch

Prefetch status

Constructors

Fresh

Prefetched

type family PrefetchResult (k :: Prefetch) where ...

Prefetched fetchers hold hashes

Equations

PrefetchResult Fresh = ()

PrefetchResult Prefetched = SHA256

Package types

type PackageName = Text

Package name, used in generating nix expr

type PackageFetcher = Version -> NixFetcher Fresh

How to create package source fetcher given a version

data Package

A package is defined with:

its name

how to track its version

how to fetch it as we have the version

Equality of packages is determined by only the name

Constructors

Package

Fields

pname :: PackageName

pversion :: VersionSource

pfetcher :: PackageFetcher