NvFetcher

Description

The main module of nvfetcher. If you want to create CLI program with it, it's enough to import only this module.

Example:

module Main where

import NvFetcher

main :: IO ()
main = runNvFetcher defaultArgs packageSet

packageSet :: PackageSet ()
packageSet = do
  define $ package "feeluown-core" fromPypi "feeluown"
  define $ package "qliveplayer" fromGitHub (IsoaSFlus, QLivePlayer)

You can find more examples of packages in Main_example.hs.

Running the created program:

main -- abbreviation of main build

main build -- build nix sources expr from given packageSet

main clean -- delete .shake dir and generated nix file

main -j -- build with parallelism

All shake options are inherited.

Documentation

data Args

Arguments for running nvfetcher

Constructors

Args

Fields

argShakeOptions :: ShakeOptions

Shake options

argTarget :: String

Build target

argBuildDir :: FilePath

Shake dir

argRules :: Rules ()

Custom rules

argActionAfterBuild :: Action ()

Action run after build rule

argActionAfterClean :: Action ()

Action run after clean rule

argRetries :: Int

Retry times

defaultArgs :: Args

Default arguments of defaultMain

Build dir is _sources.

runNvFetcher :: PackageSet () -> IO ()

Run nvfetcher with CLI options

This function calls runNvFetcherNoCLI, using Args from CLIOptions. Use this function to create your own Haskell executable program.

runNvFetcherNoCLI :: Args -> PackageSet () -> IO ()

Entry point of nvfetcher

cliOptionsToArgs :: CLIOptions -> Args

Apply CLIOptions to defaultArgs

module NvFetcher.PackageSet

module NvFetcher.Types

module NvFetcher.Types.ShakeExtras