NvFetcher.NixFetcher

Description

NixFetcher is used to describe how to fetch package sources.

There are two types of fetchers overall:

FetchGit -- nix-prefetch-git

FetchUrl -- nix-prefetch-url

As you can see the type signature of prefetch: a fetcher will be filled with the fetch result (hash) after the prefetch.

Types

data NixFetcher (k :: Prefetch)

If the package is prefetched, then we can obtain the SHA256

Constructors

FetchGit

Fields

_furl :: Text

_rev :: Version

_branch :: Maybe Text

_deepClone :: Bool

_fetchSubmodules :: Bool

_leaveDotGit :: Bool

_sha256 :: PrefetchResult k

FetchUrl

Fields

_furl :: Text

_sha256 :: PrefetchResult k

data Prefetch

Prefetch status

Constructors

Fresh

Prefetched

class ToNixExpr a where

Types can be converted into nix expr

Methods

toNixExpr :: a -> NixExpr

type family PrefetchResult (k :: Prefetch) where ...

Prefetched fetchers hold hashes

Equations

PrefetchResult Fresh = ()

PrefetchResult Prefetched = SHA256

Rules

prefetchRule :: Rules ()

Rules of nix fetcher

prefetch :: NixFetcher Fresh -> Action (NixFetcher Prefetched)

Run nix fetcher

Functions

gitHubFetcher

Arguments

:: (Text, Text)
owner and repo

-> PackageFetcher

Create a fetcher from github repo

pypiFetcher :: Text -> PackageFetcher

Create a fetcher from pypi

gitHubReleaseFetcher

Arguments

:: (Text, Text)
owner and repo

-> Text
file name

-> PackageFetcher

Create a fetcher from github release

gitFetcher :: Text -> PackageFetcher

Create a fetcher from git url

urlFetcher :: Text -> NixFetcher Fresh

Create a fetcher from url