NvFetcher.Nvchecker

Description

nvchecker is a program checking new versions of packages. We encode the checking process into shake build system, generating configuration of nvchecker and calling it externally. Now we call nvchecker for each VersionSource, which seems not to be efficient, but it's tolerable when running in parallel.

Meanwhile, we lose the capabilities of tracking version updates, i.e. normally nvchecker will help us maintain a list of old versions, so that we are able to know which package's version is updated in this run. Fortunately, we can reimplement this using shake database, see nvcheckerRule for details.

Types

data VersionSource

The input of nvchecker

Constructors

GitHubRelease

Fields

_owner :: Text

_repo :: Text

Git

Fields

_vurl :: Text

Pypi

Fields

_pypi :: Text

ArchLinux

Fields

_archpkg :: Text

Aur

Fields

_aur :: Text

Manual

Fields

_manual :: Text

Repology

Fields

_repology :: Text

_repo :: Text

data NvcheckerResult

The result of running nvchecker

Constructors

NvcheckerResult

Fields

nvNow :: Version

nvOld :: Maybe Version

nvchecker doesn't give this value, but shake restores it from last run

Rules

nvcheckerRule :: Rules ()

Rules of nvchecker

checkVersion :: VersionSource -> PackageKey -> Action NvcheckerResult

Run nvchecker