NvFetcher.Types

Description

Types used in this program.

Common types

newtype Version

Package version

Constructors

Version Text

newtype Checksum

Check sum, sha256, sri or base32, etc.

Constructors

Checksum Text

newtype ContainerDigest

Digest of a (Docker) container

Constructors

ContainerDigest Text

newtype Branch

Git branch (Nothing: master)

Constructors

Branch (Maybe Text)

type NixExpr = Text

Nix expression

data VersionChange

Version change of a package

>>> VersionChange "foo" Nothing "2.3.3"
foo: ∅ → 2.3.3

>>> VersionChange "bar" (Just "2.2.2") "2.3.3"
bar: 2.2.2 → 2.3.3

Constructors

VersionChange

Fields

vcName :: PackageName

vcOld :: Maybe Version

vcNew :: Version

newtype WithPackageKey k

Decorate a rule's key with PackageKey

Constructors

WithPackageKey (k, PackageKey)

Nvchecker types

data VersionSortMethod

Constructors

ParseVersion

Vercmp

data ListOptions

Filter-like configuration for some version sources. See https://nvchecker.readthedocs.io/en/latest/usage.html#list-options for details.

Constructors

ListOptions

Fields

_includeRegex :: Maybe Text

_excludeRegex :: Maybe Text

_sortVersionKey :: Maybe VersionSortMethod

_ignored :: Maybe Text

data VersionSource

Upstream version source for nvchecker to check

Constructors

GitHubRelease

Fields

_owner :: Text

_repo :: Text

GitHubTag

Fields

_owner :: Text

_repo :: Text

_listOptions :: ListOptions

Git

Fields

_vurl :: Text

_vbranch :: Branch

Pypi

Fields

_pypi :: Text

ArchLinux

Fields

_archpkg :: Text

Aur

Fields

_aur :: Text

Manual

Fields

_manual :: Text

Repology

Fields

_repology :: Text

_repo :: Text

Webpage

Fields

_vurl :: Text

_regex :: Text

_listOptions :: ListOptions

HttpHeader

Fields

_vurl :: Text

_regex :: Text

_listOptions :: ListOptions

OpenVsx

Fields

_ovPublisher :: Text

_ovExtName :: Text

VscodeMarketplace

Fields

_vsmPublisher :: Text

_vsmExtName :: Text

Cmd

Fields

_vcmd :: Text

Container

Fields

_vcontainer :: Text

_listOptions :: ListOptions

data NvcheckerResult

The result of nvchecker rule

Constructors

NvcheckerResult

Fields

nvNow :: Version

nvOld :: Maybe Version

last result of this nvchecker rule TODO: consider removing this field

nvStale :: Bool

stale means even nvNow comes from json file (last run) and we actually didn't run nvchecker this time. nvOld will be Nothing in this case.

data NvcheckerRaw

Parsed JSON output from nvchecker

Constructors

NvcheckerSuccess Version

NvcheckerError Text

data CheckVersion

The input of nvchecker

Constructors

CheckVersion VersionSource NvcheckerOptions

data NvcheckerOptions

Configuration available for evey version sourece. See https://nvchecker.readthedocs.io/en/latest/usage.html#global-options for details.

Constructors

NvcheckerOptions

Fields

_stripPrefix :: Maybe Text

_fromPattern :: Maybe Text

_toPattern :: Maybe Text

data UseStaleVersion

Using stale value indicates that we will NOT check for new versions if there is a known version recovered from json file or last use of the rule. Normally you don't want a stale version unless you need pin a package.

Constructors

PermanentStale
Specified in configuration file

TemporaryStale
Specified by --filter command

NoStale

Nix fetcher types

data RunFetch

The input of prefetch rule

Constructors

RunFetch ForceFetch (NixFetcher Fresh)

data ForceFetch

Whether to cache the fetched sha256

ForceFetch indicates alwaysRerun the fetcher rule

Constructors

ForceFetch

NoForceFetch

data NixFetcher (k :: FetchStatus)

If the package is prefetched, then we can obtain the SHA256

Constructors

FetchGit

Fields

_furl :: Text

_rev :: Version

_deepClone :: Bool

_fetchSubmodules :: Bool

_leaveDotGit :: Bool

_name :: Maybe Text

_sha256 :: FetchResult Checksum k

FetchGitHub

Fields

_fowner :: Text

_frepo :: Text

_rev :: Version

_deepClone :: Bool

_fetchSubmodules :: Bool

_leaveDotGit :: Bool

_name :: Maybe Text

_sha256 :: FetchResult Checksum k

FetchUrl

Fields

_furl :: Text

_name :: Maybe Text

_sha256 :: FetchResult Checksum k

FetchTarball

Fields

_furl :: Text

_sha256 :: FetchResult Checksum k

FetchDocker

Fields

_imageName :: Text

_imageTag :: Text

_imageDigest :: FetchResult ContainerDigest k

_sha256 :: FetchResult Checksum k

_fos :: Maybe Text

_farch :: Maybe Text

_finalImageName :: Maybe Text

_finalImageTag :: Maybe Text

_tlsVerify :: Maybe Bool

type family FetchResult a (k :: FetchStatus) where ...

Prefetched fetchers hold hashes

Equations

FetchResult _ Fresh = ()

FetchResult a Fetched = a

data FetchStatus

Fetch status

Constructors

Fresh

Fetched

ExtractSrc Types

data ExtractSrcQ

Extract file contents from package source e.g. Cargo.lock

Constructors

ExtractSrcQ (NixFetcher Fetched) (NonEmpty FilePath)

FetchRustGitDeps types

data FetchRustGitDepsQ

Fetch outputHashes for git dependencies in Cargo.lock. See https://github.com/NixOS/nixpkgs/blob/master/doc/languages-frameworks/rust.section.md#importing-a-cargolock-file for details. We need fetched source and the file path to Cargo.lock.

Constructors

FetchRustGitDepsQ (NixFetcher Fetched) FilePath

GetGitCommitDate types

newtype DateFormat

strftime format

Nothing defaults to %Y-%m-%d

Constructors

DateFormat (Maybe Text)

data GetGitCommitDate

Get the commit date by using shallow clone

_gformat is in. Note: Requires git >= 2.5

Constructors

GetGitCommitDate

Fields

_gurl :: Text

_grev :: Text

_gformat :: DateFormat

Core types

data Core

The key type of nvfetcher rule. See NvFetcher.Core

Constructors

Core

Package types

type PackageName = Text

Package name, used in generating nix expr

type PackageFetcher = Version -> NixFetcher Fresh

How to create package source fetcher given a version

newtype PackageExtractSrc

Constructors

PackageExtractSrc (NonEmpty FilePath)

newtype PackageCargoLockFiles

Constructors

PackageCargoLockFiles (NonEmpty FilePath)

newtype PackagePassthru

Constructors

PackagePassthru (HashMap Text Text)

data Package

A package is defined with:

its name

how to track its version

how to fetch it as we have the version

optional file paths to extract (dump to build dir)

optional Cargo.lock path (if it's a rust package)

an optional pass through map

if the package version was pinned

optional git date format (if the version source is git)

whether to always fetch a package regardless of the version changing INVARIANT: Version passed to PackageFetcher MUST be used textually, i.e. can only be concatenated with other strings, in case we can't check the equality between fetcher functions.

Constructors

Package

Fields

_pname :: PackageName

_pversion :: CheckVersion

_pfetcher :: PackageFetcher

_pextract :: Maybe PackageExtractSrc

_pcargo :: Maybe PackageCargoLockFiles

_ppassthru :: PackagePassthru

_ppinned :: UseStaleVersion

_pgitdateformat :: DateFormat

_pforcefetch :: ForceFetch

newtype PackageKey

Package key is the name of a package. We use this type to index packages.

Constructors

PackageKey PackageName

data PackageResult

Result type of Core

Constructors

PackageResult

Fields

_prname :: PackageName

_prversion :: NvcheckerResult

_prfetched :: NixFetcher 'Fetched

_prpassthru :: Maybe (HashMap Text Text)

_prextract :: Maybe (HashMap FilePath NixExpr)

extracted file name -> file path in build dir

_prcargolock :: Maybe (HashMap FilePath (NixExpr, HashMap Text Checksum))

cargo lock file path in build dir -> (file path in nix, git dependencies)

_prpinned :: UseStaleVersion

_prgitdate :: Maybe Text