NvFetcher.Types

Description

Types used in this program.

Common types

newtype Version

Package version

Constructors

Version Text

newtype SHA256

SHA 256 sum

Constructors

SHA256 Text

type NixExpr = Text

Nix expression

data VersionChange

Version change of a package

>>> VersionChange "foo" Nothing "2.3.3"
foo: ∅ → 2.3.3

>>> VersionChange "bar" (Just "2.2.2") "2.3.3"
bar: 2.2.2 → 2.3.3

Constructors

VersionChange

Fields

vcName :: PackageName

vcOld :: Maybe Version

vcNew :: Version

newtype WithPackageKey k

Decorate a rule's key with PackageKey

Constructors

WithPackageKey (k, PackageKey)

data Core

The key type of nvfetcher rule. See NvFetcher.Core

Constructors

Core

Nvchecker types

data VersionSource

The input of nvchecker

Constructors

GitHubRelease

Fields

_owner :: Text

_repo :: Text

Git

Fields

_vurl :: Text

Pypi

Fields

_pypi :: Text

ArchLinux

Fields

_archpkg :: Text

Aur

Fields

_aur :: Text

Manual

Fields

_manual :: Text

Repology

Fields

_repology :: Text

_repo :: Text

data NvcheckerResult

The result of running nvchecker

Constructors

NvcheckerResult

Fields

nvNow :: Version

nvOld :: Maybe Version

nvchecker doesn't give this value, but shake restores it from last run

Nix fetcher types

data NixFetcher (k :: Prefetch)

If the package is prefetched, then we can obtain the SHA256

Constructors

FetchGit

Fields

_furl :: Text

_rev :: Version

_branch :: Maybe Text

_deepClone :: Bool

_fetchSubmodules :: Bool

_leaveDotGit :: Bool

_sha256 :: PrefetchResult k

FetchUrl

Fields

_furl :: Text

_sha256 :: PrefetchResult k

data Prefetch

Prefetch status

Constructors

Fresh

Prefetched

type family PrefetchResult (k :: Prefetch) where ...

Prefetched fetchers hold hashes

Equations

PrefetchResult Fresh = ()

PrefetchResult Prefetched = SHA256

Package types

type PackageName = Text

Package name, used in generating nix expr

type PackageFetcher = Version -> NixFetcher Fresh

How to create package source fetcher given a version

data Package

A package is defined with:

its name

how to track its version

how to fetch it as we have the version

INVARIANT: Version passed to PackageFetcher MUST be used textually, i.e. can only be concatenated with other strings, in case we can't check the equality between fetcher functions.

Constructors

Package

Fields

_pname :: PackageName

_pversion :: VersionSource

_pfetcher :: PackageFetcher

newtype PackageKey

Package key is the name of a package. We use this type to index packages.

Constructors

PackageKey PackageName